+++ packages/figma/plugin-src/toUIMix/SVGLikeNodeChecker.ts
