+++ packages/figma/plugin-src/toUIMix/node.ts
@@ -1,0 +1,138 @@
+import { createId } from "@paralleldrive/cuid2";
+import { compact } from "../util";
+import { StyleJSON, NodeJSON } from "@uimix/node-data";
+import { svgLikeNodeChecker } from "./SVGLikeNodeChecker";
+import {
+  getFrameStyle,
+  getGroupStyle,
+  getImageStyle,
+  getTextStyle,
+} from "./style";
+
+function isSingleImageFill(
+  fills: MinimalFillsMixin["fills"]
+): fills is readonly [ImagePaint] {
+  return (
+    fills !== figma.mixed && fills.length === 1 && fills[0].type === "IMAGE"
+  );
+}
+
+interface NodeWithStyle extends Omit<NodeJSON, "index" | "parent"> {
+  id: string;
+  style: Partial<StyleJSON>;
+  children: NodeWithStyle[];
+}
+
+async function figmaToMacaron(
+  node: SceneNode,
+  parentLayout: BaseFrameMixin["layoutMode"],
+  offset: [number, number]
+): Promise<NodeWithStyle | undefined> {
+  // FigJam nodes are not supported
+  if (
+    node.type === "STICKY" ||
+    node.type === "CONNECTOR" ||
+    node.type === "SHAPE_WITH_TEXT" ||
+    node.type === "CODE_BLOCK" ||
+    node.type === "STAMP" ||
+    node.type === "WIDGET" ||
+    node.type === "EMBED" ||
+    node.type === "LINK_UNFURL" ||
+    node.type === "MEDIA" ||
+    node.type === "WASHI_TAPE" ||
+    node.type === "SECTION" ||
+    node.type === "TABLE"
+  ) {
+    return;
+  }
+
+  if (
+    node.type === "RECTANGLE" &&
+    isSingleImageFill(node.fills) &&
+    node.fills[0].imageHash
+  ) {
+    return {
+      id: createId(),
+      type: "image",
+      name: node.name,
+      style: await getImageStyle(node, parentLayout, offset),
+      children: [],
+    };
+  }
+
+  if (svgLikeNodeChecker.check(node)) {
+    try {
+      const svg = await node.exportAsync({ format: "SVG" });
+      const svgText = String.fromCharCode(...svg);
+
+      return {
+        id: createId(),
+        type: "image",
+        name: node.name,
+        style: {
+          ...getPositionStyle(node, parentLayout, offset),
+          width: { type: "hugContents" },
+          height: { type: "hugContents" },
+          image: {
+            type: "svg",
+            content: svgText,
+          },
+        },
+      };
+    } catch (error) {
+      console.error(`error exporting ${node.name} to SVG`);
+      console.error(String(error));
+    }
+  }
+
+  switch (node.type) {
+    case "GROUP": {
+      return {
+        id: createId(),
+        type: "frame",
+        name: node.name,
+        style: await getGroupStyle(node, parentLayout, offset),
+        children: await figmaNodesToMacaron(node.children, "NONE", [
+          node.x,
+          node.y,
+        ]),
+      };
+    }
+    case "COMPONENT":
+    case "COMPONENT_SET":
+    case "INSTANCE":
+    case "FRAME": {
+      return {
+        id: createId(),
+        type: "frame",
+        name: node.name,
+        style: await getFrameStyle(node, parentLayout, offset),
+        children: await figmaNodesToMacaron(node.children, node.layoutMode, [
+          node.strokeLeftWeight,
+          node.strokeTopWeight,
+        ]),
+      };
+    }
+    case "TEXT": {
+      return {
+        id: createId(),
+        type: "text",
+        name: node.name,
+        style: await getTextStyle(node, parentLayout, offset),
+        children: [],
+      };
+    }
+  }
+}
+
+export async function figmaNodesToMacaron(
+  nodes: readonly SceneNode[],
+  parentLayout: BaseFrameMixin["layoutMode"],
+  offset: [number, number]
+): Promise<NodeWithStyle[]> {
+  return compact(
+    await Promise.all(
+      nodes.map((child) => figmaToMacaron(child, parentLayout, offset))
+    )
+  );
+}
